ecourses/requirements.txt
amqp==5.2.0
app==0.0.1
asgiref==3.7.2
billiard==3.6.4.0
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
celery==5.0.5
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.37.0
colorama==0.4.6
cryptography==42.0.2
decorators==2.0.7
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.6
django-allauth==0.60.1
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-AdminLTE3==1.0.9
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
stripe==7.10.0
twilio==6.0.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.0.1
